fix: resolve assorted training and inference errors
- Fixed SyntaxError due to duplicate keyword arguments in `train.py`.
- Corrected typecast missing in `losses/hybrid_loss.py` to prevent FP16 Mixed Precision (AMP) SSIM failures.
- Updated initialization in `test_airnet.py` to correctly drop an unsupported argument `use_lpips=False`.
- Made `datasets/kla_dataset.py` robust against missing sample count errors.
- Excluded outputs and pycache directories from git using `.gitignore`.

+++ losses/hybrid_loss.py
@@ -286,7 +286,7 @@ class FallbackSSIM(nn.Module):
         sigma12 = F.conv2d(img1_f * img2_f, window, padding=self.window_size // 2, groups=self.channel) - mu1_mu2
 
         ssim_map = ((2 * mu1_mu2 + c1) * (2 * sigma12 + c2)) / ((mu1_sq + mu2_sq + c1) * (sigma1_sq + sigma2_sq + c2))
-        return ssim_map.mean()
+        return ssim_map.mean().to(img1.dtype)
 
 
 HybridLoss = AIRNetHybridLoss
